+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 # neg - CHANGELOG.md
 
 
+## neg - 1.2.0    released 2013-02-28
+
+- implement Parser.recursive?
+- simplify Parser.to_s
+
+
 ## neg - 1.1.0    released 2013-02-28
 
 - giving the "right" error on unconsumed input